Extract message-names and GhError into submodules to clear file-size gate
The close-reason additions pushed src/messages.rs and src/github/mod.rs over
the 1000-line hard limit when merged with the latest main. Move the
message_names! macro and its invocations into src/messages/names.rs, and the
GhError enum + impls into src/github/error.rs. Both are cohesive, self-
contained concerns; the extraction preserves all public APIs via re-exports.
